feat: moved search component in code-editor
