+++ packages/mantine/src/components/code-editor/search/index.ts
